+++ src/app/app.module.ts
@@ -22,8 +22,6 @@ import { CuadraFacturacionComponent } from './components/cuadra-facturacion/cuad
 import { CuadraFacturacionGenerarComponent } from './components/cuadra-facturacion-generar/cuadra-facturacion-generar.component';
 import { MywizardRvComponent } from './components/mywizard-rv/mywizard-rv.component';
 import { MywizardRvGeneracionComponent } from './components/mywizard-rv-generacion/mywizard-rv-generacion.component';
-import { FechasComponent } from './components/fechas/fechas.component';
-import { FechasGeneracionComponent } from './components/fechas-generacion/fechas-generacion.component';
 
 import { GenerarArsGplanComponent } from './components/generar-ars-gplan/generar-ars-gplan.component';
 import { VerArsJiraPlanComponent } from './components/ver-ars-jira-plan/ver-ars-jira-plan.component';
@@ -35,9 +33,6 @@ import { MyteMmeGeneracionComponent } from './components/mytemme/myte-mme-genera
 
 import { MetricasAMModule } from './metricas-am/metricas-am.module';
 import { MantenimientoModule } from './mantenimiento/mantenimiento.module';
-
-import { FechaDMSPipe } from 'src/app/pipes/fecha-dms.pipe';
-
 import { FechasDMSModule } from './fechasDMS/fechasDMS.module';
 import { ValidarHHModule } from './validarHH/validarHH.module';
 import { SharedModule } from './shared/shared.module';
@@ -49,7 +44,6 @@ import { SharedModule } from './shared/shared.module';
     NavbarComponent,
     GenerarInformeComponent,
     FechasPipe,
-    FechaDMSPipe,
     GenerarCapacityComponent,
     VerCapacityComponent,
     GenerarArsGfacComponent,
@@ -61,8 +55,6 @@ import { SharedModule } from './shared/shared.module';
     CuadraFacturacionGenerarComponent,
     MywizardRvComponent,
     MywizardRvGeneracionComponent,
-    FechasComponent,
-    FechasGeneracionComponent,
     GenerarArsGplanComponent,
     VerArsJiraPlanComponent,
     MyteMmeGeneracionComponent,
