+++ src/app/components/navbar/navbar.component.html
@@ -46,15 +46,12 @@
                     <i class="far fa-calendar-alt"></i>&nbsp; Fechas DMS
                 </a>
                 <div class="dropdown-menu" aria-labelledby="navbarDropdownFechasDMS">
-                  
-                    <a class="dropdown-item" href="#" [routerLink]="['fechas']" routerLinkActive="active">Valida DMS</a>
-
                     <a 
                         class="dropdown-item" 
                         href="#" 
                         [routerLink]="['fechasDMS/validar/formulario']" 
                         routerLinkActive="active">
-                        Valida DMS test
+                        Valida DMS
                     </a>
 
                     <a 
@@ -64,7 +61,6 @@
                         routerLinkActive="active">
                             Comparar
                     </a>
-
                 </div>
             </li>
