+++ src/app/fechasDMS/validar/components/formulario/formulario.component.html
@@ -1,1 +1,42 @@
-<p>formulario works!</p>
+<div class="container">
+    <div class="row justify-content-center">
+        <div class="col-12 col-sm-6 ">
+            <div class="row">
+                <div class="col ">
+                    <h4>Subir Archivos</h4>
+                </div>
+            </div>
+            <form autocomplete="off" [formGroup]="formulario" (ngSubmit)="guardar()">
+
+                <div class="row my-3">
+                    <div class="col-12 col-sm-4">
+                        Fecha:
+                    </div>
+                    <div class="col-12 col-sm-8">
+                        <input [class.is-invalid]="fechaNoValido" formControlName="fecha" type="month" class="date form-control" (change)="cambiarFecha($event)"  min="{{fechaMin}}" max="{{fechaMax}}" >
+                        <small *ngIf="fechaNoValido" class="text-danger">Seleccione Fecha</small>
+                    </div>
+                </div>
+ 
+                <div class="row my-3">
+                    <div class="col-12 col-sm-4">
+                        <i *ngIf="estadoDms==1" class="fas fa-chevron-right text-dark"></i><i *ngIf="estadoDms==2" class="fas fa-spinner fa-spin"></i><i *ngIf="estadoDms==3" class="fas fa-check text-success"></i><i *ngIf="estadoDms==4" class="fas fa-times text-danger"></i>&nbsp;
+                        DMS:
+                    </div>
+                    <div class="col-12 col-sm-8">
+                        <input [class.is-invalid]="dmsNoValido" formControlName="dms" type="file" class="form-control-file" (change)="uploadDms($event)" (click)="click(1)">
+                        <small *ngIf="dmsNoValido" class="text-danger">Seleccione Archivo<br/></small>
+                        <small>ARS -> Reportes -> DMS (solo evolutivo mayor)</small>
+                    </div>
+                </div>
+           
+                <div class="row my-3 mt-5">
+                    <div class="col">
+                        <button class="btn btn-primary btn-block" type="submit">Generar</button>
+                    </div>
+                </div>
+            </form>
+
+        </div>
+    </div>
+</div>
